django-rest-api/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.1
django-cors-headers>=4.3.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.61.1
django-rest-auth==0.9.5
django-filter==24.1
Pillow>=10.0.0
requests==2.31.0
core==0.1.0
numpy>=1.24.0
opencv-python>=4.8.0
pytesseract>=0.3.10
scikit-learn>=1.3.0
matplotlib>=3.7.0
ultralytics>=8.0.0
openai>=1.0.0
xgboost>=2.0.0